Broken: Stack overflow. Trying to subdivide triangles in 3D is bad
I realized that it's hard to subdivide the triangles in 3D, since
you can't just take the midpoint on the edge and normalize. That
will "cut corners" near the poles and if the two longitudes are
> PI apart. That will make the triangle go on the wrong side of the
planet.

I will now retry subdividing in 2D before mapping on a sphere... again